requirements.txt
# TranscriMatic Requirements
# Audio Transcription & Analysis System

# Core Dependencies
pyyaml>=6.0.1
pathlib2>=2.3.7

# USB Device Monitoring
pyudev>=0.24.0  # Linux only - see requirements-macos.txt for macOS

# Audio Processing
pydub>=0.25.1
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.1
noisereduce>=3.0.0

# Transcription
faster-whisper>=0.9.0
torch>=2.0.0
pyannote.audio>=3.0.0

# Natural Language Processing
spacy>=3.6.0
scikit-learn>=1.3.0
nltk>=3.8.0

# LLM Providers
requests>=2.31.0
google-generativeai>=0.3.0
openai>=1.3.0
backoff>=2.2.1
langchain>=0.0.340  # Optional

# Output Formatting
jinja2>=3.1.2
markdown>=3.4.4

# Database
# SQLite3 is included in Python stdlib

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Optional Performance Enhancements
# nvidia-cublas-cu11>=11.10.3.66  # For CUDA support
# nvidia-cudnn-cu11>=8.5.0.96     # For CUDA support